+++ e2e/05-max-warnings/contracts/Foo.sol
@@ -1,5 +1,5 @@
 // SPDX-License-Identifier: MIT
-pragma solidity >=0.5.8;
+pragma solidity >=0.8.0;
 
 contract Foo {
     uint256 public constant test1 = 1;