+++ test/rules/security/compiler-version.js
@@ -2,6 +2,8 @@ const assert = require('assert')
 const { assertNoErrors, assertErrorCount, assertErrorMessage } = require('../../common/asserts')
 const linter = require('../../../lib/index')
 
+const DEFAULT_SEMVER = '^0.8.0'
+
 describe('Linter - compiler-version', () => {
   it('should disable only one compiler error on next line', () => {
     const report = linter.processStr(
@@ -11,7 +13,7 @@ describe('Linter - compiler-version', () => {
                 pragma solidity 0.3.4;
             `,
       {
-        rules: { 'compiler-version': ['error', '^0.5.2'] },
+        rules: { 'compiler-version': ['error', DEFAULT_SEMVER] },
       }
     )
 
@@ -26,7 +28,7 @@ describe('Linter - compiler-version', () => {
               pragma solidity 0.3.4;
           `,
       {
-        rules: { 'compiler-version': ['error', '^0.5.2'] },
+        rules: { 'compiler-version': ['error', DEFAULT_SEMVER] },
       }
     )
 
@@ -41,7 +43,7 @@ describe('Linter - compiler-version', () => {
                 pragma solidity 0.3.4;
             `,
       {
-        rules: { 'compiler-version': ['error', '^0.5.2'] },
+        rules: { 'compiler-version': ['error', DEFAULT_SEMVER] },
       }
     )
 
@@ -56,7 +58,7 @@ describe('Linter - compiler-version', () => {
                 pragma solidity 0.3.4;
             `,
       {
-        rules: { 'compiler-version': ['error', '^0.5.2'] },
+        rules: { 'compiler-version': ['error', DEFAULT_SEMVER] },
       }
     )
 
@@ -72,12 +74,12 @@ describe('Linter - compiler-version', () => {
                 pragma solidity 0.3.4;
             `,
       {
-        rules: { 'compiler-version': ['error', '^0.5.2'] },
+        rules: { 'compiler-version': ['error', DEFAULT_SEMVER] },
       }
     )
 
     assertErrorCount(report, 1)
-    assertErrorMessage(report, '0.5.2')
+    assertErrorMessage(report, DEFAULT_SEMVER)
   })
 
   it('should disable all errors', () => {
@@ -88,7 +90,7 @@ describe('Linter - compiler-version', () => {
                 pragma solidity 0.3.4;
             `,
       {
-        rules: { 'compiler-version': ['error', '^0.5.2'] },
+        rules: { 'compiler-version': ['error', DEFAULT_SEMVER] },
       }
     )
     assertNoErrors(report)
@@ -103,58 +105,58 @@ describe('Linter - compiler-version', () => {
                 pragma solidity ^0.4.4;
             `,
       {
-        rules: { 'compiler-version': ['error', '^0.5.2'] },
+        rules: { 'compiler-version': ['error', DEFAULT_SEMVER] },
       }
     )
 
     assertErrorCount(report, 1)
-    assertErrorMessage(report, '0.5.2')
+    assertErrorMessage(report, DEFAULT_SEMVER)
   })
 
   it('should return compiler version error', () => {
     const report = linter.processStr('pragma solidity 0.3.4;', {
-      rules: { 'compiler-version': ['error', '^0.5.2'] },
+      rules: { 'compiler-version': ['error', DEFAULT_SEMVER] },
     })
 
     assert.equal(report.errorCount, 1)
-    assert.ok(report.reports[0].message.includes('0.5.2'))
+    assert.ok(report.reports[0].message.includes(DEFAULT_SEMVER))
   })
 
   it('should not report compiler version error on exact match', () => {
-    const report = linter.processStr('pragma solidity 0.5.2;', {
-      rules: { 'compiler-version': ['error', '0.5.2'] },
+    const report = linter.processStr('pragma solidity 0.8.0;', {
+      rules: { 'compiler-version': ['error', '0.8.0'] },
     })
 
     assert.equal(report.errorCount, 0)
   })
 
   it('should not report compiler version error on range match', () => {
-    const report = linter.processStr('pragma solidity ^0.5.2;', {
-      rules: { 'compiler-version': ['error', '^0.5.2'] },
+    const report = linter.processStr('pragma solidity ^0.8.0;', {
+      rules: { 'compiler-version': ['error', DEFAULT_SEMVER] },
     })
 
     assert.equal(report.errorCount, 0)
   })
 
   it('should not report compiler version error on patch bump', () => {
-    const report = linter.processStr('pragma solidity 0.5.3;', {
-      rules: { 'compiler-version': ['error', '^0.5.2'] },
+    const report = linter.processStr('pragma solidity 0.8.1;', {
+      rules: { 'compiler-version': ['error', DEFAULT_SEMVER] },
     })
 
     assert.equal(report.errorCount, 0)
   })
 
   it('should not report compiler version error on range match', () => {
-    const report = linter.processStr('pragma solidity ^0.5.3;', {
-      rules: { 'compiler-version': ['error', '^0.5.2'] },
+    const report = linter.processStr('pragma solidity ^0.8.2;', {
+      rules: { 'compiler-version': ['error', DEFAULT_SEMVER] },
     })
 
     assert.equal(report.errorCount, 0)
   })
 
   it('should report compiler version error on range not matching', () => {
-    const report = linter.processStr('pragma solidity ^0.5.2;', {
-      rules: { 'compiler-version': ['error', '^0.5.3'] },
+    const report = linter.processStr('pragma solidity ^0.8.1;', {
+      rules: { 'compiler-version': ['error', '^0.8.3'] },
     })
 
     assert.equal(report.errorCount, 1)
@@ -162,7 +164,7 @@ describe('Linter - compiler-version', () => {
 
   it('should report compiler version error on minor bump', () => {
     const report = linter.processStr('pragma solidity 0.6.0;', {
-      rules: { 'compiler-version': ['error', '^0.5.2'] },
+      rules: { 'compiler-version': ['error', DEFAULT_SEMVER] },
     })
 
     assert.equal(report.errorCount, 1)
@@ -170,7 +172,7 @@ describe('Linter - compiler-version', () => {
 
   it(`should report compiler version error if pragma doesn't exist`, () => {
     const report = linter.processStr('contract Foo {}', {
-      rules: { 'compiler-version': ['error', '^0.5.2'] },
+      rules: { 'compiler-version': ['error', DEFAULT_SEMVER] },
     })
 
     assert.equal(report.errorCount, 1)
@@ -178,10 +180,10 @@ describe('Linter - compiler-version', () => {
 
   it(`should not report compiler version error if pragma exist`, () => {
     const report = linter.processStr(
-      `pragma solidity 0.6.0;
+      `pragma solidity 0.8.2;
       contract Foo {}`,
       {
-        rules: { 'compiler-version': ['error', '^0.6.0'] },
+        rules: { 'compiler-version': ['error', '^0.8.2'] },
       }
     )
 
@@ -201,5 +203,63 @@ describe('Linter - compiler-version', () => {
     )
 
     assert.equal(report.errorCount, 0)
+  })
+
+  it(`should not report compiler version error using default and correct pragma`, () => {
+    const report = linter.processStr(
+      `pragma solidity ^0.8.1;
+       pragma experimental ABIEncoderV2;
+       
+       contract Main {
+       }`,
+      {
+        rules: { 'compiler-version': 'error' },
+      }
+    )
+
+    assert.equal(report.errorCount, 0)
+  })
+
+  it(`should report compiler version error using default and lower pragma`, () => {
+    const report = linter.processStr(
+      `pragma solidity ^0.7.4;
+       
+       contract Main {
+       }`,
+      {
+        rules: { 'compiler-version': 'error' },
+      }
+    )
+
+    assert.equal(report.errorCount, 1)
+    assertErrorMessage(report, DEFAULT_SEMVER)
+  })
+
+  it(`should not report compiler version error using >= and default and correct pragma`, () => {
+    const report = linter.processStr(
+      `pragma solidity >=0.8.0;
+       
+       contract Main {
+       }`,
+      {
+        rules: { 'compiler-version': 'error' },
+      }
+    )
+
+    assert.equal(report.errorCount, 0)
+  })
+
+  it(`should report compiler version error using >= and default and lower pragma`, () => {
+    const report = linter.processStr(
+      `pragma solidity >=0.7.4;
+       
+       contract Main {
+       }`,
+      {
+        rules: { 'compiler-version': 'error' },
+      }
+    )
+    assert.equal(report.errorCount, 1)
+    assertErrorMessage(report, DEFAULT_SEMVER)
   })
 })